ChattingRoomRepository를 ChatRoomService로 리팩토링
