+++ src/main/java/com/example/modiraa/controller/ChatController.java
@@ -2,7 +2,7 @@ package com.example.modiraa.controller;
 
 import com.example.modiraa.config.jwt.JwtAuthorizationFilter;
 import com.example.modiraa.model.ChatMessage;
-import com.example.modiraa.repository.ChattingRoomRepository;
+import com.example.modiraa.service.ChatRoomService;
 import com.example.modiraa.service.ChatService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -16,7 +16,7 @@ import org.springframework.stereotype.Controller;
 public class ChatController {
 
     private final JwtAuthorizationFilter jwtAuthorizationFilter;
-    private final ChattingRoomRepository chattingRoomRepository;
+    private final ChatRoomService chatRoomService;
     private final ChatService chatService;
 
     // websocket "/pub/chat/message"로 들어오는 메시징을 처리한다.
@@ -28,7 +28,7 @@ public class ChatController {
         // 로그인 회원 정보로 대화명 설정
         message.setSender(nickname);
         // 채팅방 인원수 세팅
-        message.setUserCount(chattingRoomRepository.getUserCount(message.getRoomId()));
+        message.setUserCount(chatRoomService.getUserCount(message.getRoomId()));
 
         // Websocket에 발행된 메시지를 redis로 발행(publish)
         chatService.sendChatMessage(message);
