+++ src/main/java/com/example/modiraa/service/ChatService.java
@@ -1,7 +1,6 @@
 package com.example.modiraa.service;
 
 import com.example.modiraa.model.ChatMessage;
-import com.example.modiraa.repository.ChattingRoomRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.redis.core.RedisTemplate;
@@ -15,7 +14,7 @@ public class ChatService {
 
     private final ChannelTopic channelTopic;
     private final RedisTemplate redisTemplate;
-    private final ChattingRoomRepository chattingRoomRepository;
+    private final ChatRoomService chatRoomService;
 
     // destination정보에서 roomId 추출
     public String getRoomId(String destination) {
@@ -28,7 +27,7 @@ public class ChatService {
 
     // 채팅방에 메시지 발송
     public void sendChatMessage(ChatMessage chatMessage) {
-        chatMessage.setUserCount(chattingRoomRepository.getUserCount(chatMessage.getRoomId()));
+        chatMessage.setUserCount(chatRoomService.getUserCount(chatMessage.getRoomId()));
         if (ChatMessage.MessageType.ENTER.equals(chatMessage.getType())) {
             chatMessage.setMessage(chatMessage.getSender() + "님이 방에 입장했습니다.");
             //chatMessage.setSender("[알림]");
