+++ src/main/resources/static/roomdetail.html
@@ -75,7 +75,7 @@
                 _this.sendMessage('ENTER');
             }, function(error) {
                 alert("서버 연결에 실패 하였습니다. 다시 접속해 주십시요.");
-                location.href="/chat/room";
+                location.href="room.html";
             });
         },
         methods: {
